--- a/Assets/AddressableAssetsData/link.xml.meta
+++ b/Assets/AddressableAssetsData/link.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 72085dd1d7cdde64a8724933110057a1
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: