--- a/Assets/AddressableAssetsData/link.xml.meta
+++ b/Assets/AddressableAssetsData/link.xml.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0be145e73194b1e49b4bee2c5511c362
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: