--- a/Assets/AddressableAssetsData/link.xml.meta
+++ b/Assets/AddressableAssetsData/link.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 874f216e8f1019b40b7a01548c202076
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: