--- a/Assets/Boats/Scripts/SpeedboatScript.cs
+++ b/Assets/Boats/Scripts/SpeedboatScript.cs
@@ -6,18 +6,14 @@ using UnityEngine;
 public class SpeedboatScript : MonoBehaviour
 {
     // Start is called before the first frame update
-    private GameObject advancedOcean;
-    private AdvancedOcean advancedOceanScript;
     void Start()
     {
-        advancedOcean = GameObject.Find("Ocean");
-        advancedOceanScript = advancedOcean.GetComponent<AdvancedOcean>();
     }
 
     // Update is called once per frame
     void Update()
     {
-        transform.position = new Vector3(transform.position.x, advancedOceanScript.GetWaterHeight(transform.position) *0.95f - 0.05f, transform.position.z);
+        transform.position = new Vector3(transform.position.x, transform.position.y + (Mathf.Sin(Time.time) * 0.0005f), transform.position.z);
         //new Vector3(transform.position.x, advancedOceanScript.GetWaterHeight(transform.position), transform.position.z);
         //new Vector3(transform.position.x, advancedOceanScript.GetWaterHeight(new Vector3(transform.position.x * 16, transform.position.y, transform.position.z * 16))); 
     }