--- a/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
+++ b/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
@@ -279,6 +279,26 @@ PrefabInstance:
       value: 1
       objectReference: {fileID: 0}
     - target: {fileID: 13700002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
+      propertyPath: m_Quality
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 13700002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
+      propertyPath: m_CastShadows
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 13700002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
+      propertyPath: m_ReceiveShadows
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 13700002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
+      propertyPath: m_LightProbeUsage
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 13700002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
+      propertyPath: m_ReflectionProbeUsage
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 13700002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Materials.Array.data[0]
       value: 
       objectReference: {fileID: 2100000, guid: ff8e972ebe73bdd458bd32654d03ddc1, type: 2}