--- a/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
+++ b/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
@@ -10,7 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4545945066944266035}
   - component: {fileID: 4545945066944266036}
-  m_Layer: 4
+  m_Layer: 6
   m_Name: FishSprite(Clone)
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -95,7 +95,7 @@ GameObject:
   - component: {fileID: 5081597436798597803}
   - component: {fileID: 5081597436798597802}
   - component: {fileID: 4545945065992470969}
-  m_Layer: 4
+  m_Layer: 6
   m_Name: Fish
   m_TagString: Fish
   m_Icon: {fileID: 0}
@@ -168,51 +168,51 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 100000, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100002, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100004, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100006, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100008, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100010, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100012, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100014, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100016, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100018, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100020, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 100022, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_Layer
-      value: 4
+      value: 6
       objectReference: {fileID: 0}
     - target: {fileID: 400000, guid: d5b11c64f880d5e48a34b0d9b479efac, type: 3}
       propertyPath: m_RootOrder