--- a/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
+++ b/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
@@ -74,7 +74,7 @@ SpriteRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300000, guid: d2143803b5c848945845fd7056505ea3, type: 3}
-  m_Color: {r: 0.8301887, g: 0.8301887, b: 0.8301887, a: 1}
+  m_Color: {r: 1, g: 1, b: 1, a: 0.39607844}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
@@ -146,7 +146,7 @@ LODGroup:
   m_AnimateCrossFading: 0
   m_LastLODIsBillboard: 0
   m_LODs:
-  - screenRelativeHeight: 0.42233375
+  - screenRelativeHeight: 0.3705193
     fadeTransitionWidth: 0.453
     renderers:
     - renderer: {fileID: 8074297095499593951}