--- a/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
+++ b/Assets/FishFeeding/Components/Fish/Prefabs/Fish.prefab
@@ -16,7 +16,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &4545945066944266035
 Transform:
   m_ObjectHideFlags: 0
@@ -140,20 +140,16 @@ LODGroup:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5081597436798597801}
   serializedVersion: 2
-  m_LocalReferencePoint: {x: -0.030149132, y: -0.0122906715, z: -0.14157656}
-  m_Size: 0.8931647
+  m_LocalReferencePoint: {x: -0.030149147, y: -0.0122906715, z: -0.14297438}
+  m_Size: 0.89036906
   m_FadeMode: 0
   m_AnimateCrossFading: 0
   m_LastLODIsBillboard: 0
   m_LODs:
-  - screenRelativeHeight: 0.3705193
+  - screenRelativeHeight: 0.20278774
     fadeTransitionWidth: 0.453
     renderers:
     - renderer: {fileID: 8074297095499593951}
-  - screenRelativeHeight: 0.17700364
-    fadeTransitionWidth: 0.452
-    renderers:
-    - renderer: {fileID: 4545945066944266036}
   m_Enabled: 1
 --- !u!1001 &8074297095511192445
 PrefabInstance: