--- a/Assets/FishFeeding/Components/Fish/Scripts/FishScript.cs
+++ b/Assets/FishFeeding/Components/Fish/Scripts/FishScript.cs
@@ -25,6 +25,7 @@ public class FishScript : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
+        Random.InitState(System.DateTime.Now.Millisecond);
         InvokeRepeating(nameof(PeriodicUpdates), Random.Range(0.0f, 5.0f), 3.0f);
         fishSystem = gameObject.transform.parent.gameObject;
         fishSystemPosition = fishSystem.transform.position;