--- a/Assets/FishFeeding/Components/Fish/Scripts/FishScript.cs
+++ b/Assets/FishFeeding/Components/Fish/Scripts/FishScript.cs
@@ -124,8 +124,6 @@ public class FishScript : MonoBehaviour
      */   
     void PeriodicUpdates()
     {
-        //fishSystemPosition = fishSystem.transform.position;
-
         fishPosition = gameObject.transform.position;
         float posY = fishPosition.y;