--- a/Assets/FishFeeding/Components/FishSystem/Prefabs/FishSystem.prefab
+++ b/Assets/FishFeeding/Components/FishSystem/Prefabs/FishSystem.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 1741675464848247204}
   - component: {fileID: 4079745591372322438}
   - component: {fileID: 4079745591372322437}
-  m_Layer: 0
+  m_Layer: 7
   m_Name: FishSystem
   m_TagString: Fish System
   m_Icon: {fileID: 0}