--- a/Assets/FishFeeding/Components/FishSystem/Prefabs/FishSystem.prefab
+++ b/Assets/FishFeeding/Components/FishSystem/Prefabs/FishSystem.prefab
@@ -32,6 +32,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 3215075958956321798}
+  - {fileID: 5907439905427248957}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -4841,6 +4842,111 @@ ParticleSystemRenderer:
   m_MeshWeighting2: 1
   m_MeshWeighting3: 1
   m_MaskInteraction: 0
+--- !u!1001 &4152430188330870484
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 3402545839537747527}
+    m_Modifications:
+    - target: {fileID: 4961586719280037294, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_Name
+      value: MerdCamera
+      objectReference: {fileID: 0}
+    - target: {fileID: 5569876810218966620, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7117804689071527273, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_Size.y
+      value: 0.822676
+      objectReference: {fileID: 0}
+    - target: {fileID: 7117804689071527273, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_Size.z
+      value: 0.49072742
+      objectReference: {fileID: 0}
+    - target: {fileID: 7117804689071527273, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_Center.y
+      value: 0.035389375
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalScale.y
+      value: 0.125
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalScale.z
+      value: 0.2
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 5fb391782e764de4b8570d09bc5a7b9d, type: 3}
+--- !u!4 &5907439905427248957 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 7519626040250249705, guid: 5fb391782e764de4b8570d09bc5a7b9d,
+    type: 3}
+  m_PrefabInstance: {fileID: 4152430188330870484}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &7988733956785495422
 PrefabInstance:
   m_ObjectHideFlags: 0