--- a/Assets/FishFeeding/Components/MerdCamera/Materials/MerdCameraTexture.renderTexture
+++ b/Assets/FishFeeding/Components/MerdCamera/Materials/MerdCameraTexture.renderTexture
@@ -14,11 +14,11 @@ RenderTexture:
   m_DownscaleFallback: 0
   m_IsAlphaChannelOptional: 0
   serializedVersion: 5
-  m_Width: 256
+  m_Width: 401
   m_Height: 256
   m_AntiAliasing: 1
   m_MipCount: -1
-  m_DepthStencilFormat: 0
+  m_DepthStencilFormat: 90
   m_ColorFormat: 8
   m_MipMap: 1
   m_GenerateMips: 1