--- a/Assets/FishFeeding/Components/MerdCamera/Materials/UnderwaterShade.mat.meta
+++ b/Assets/FishFeeding/Components/MerdCamera/Materials/UnderwaterShade.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e4c0c8a4c4b366f4286fd9e3f8c03268
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: