--- a/Assets/FishFeeding/Components/MerdCamera/Prefabs/MerdCamera.prefab
+++ b/Assets/FishFeeding/Components/MerdCamera/Prefabs/MerdCamera.prefab
@@ -57,7 +57,7 @@ Camera:
     width: 1
     height: 1
   near clip plane: 0.3
-  far clip plane: 90
+  far clip plane: 80
   field of view: 60
   orthographic: 0
   orthographic size: 5