<!--- Not useful for our project for the moment, but may be used later
* **Please check if the PR fulfills these requirements**
- [ ] The commit message follows our guidelines
- [ ] Tests for the changes have been added (for bug fixes/features)
- [ ] Docs have been added / updated (for bug fixes / features)
--->
* **What kind of change does this PR introduce?**
This greatly improves performance in the fish feeding scenario in builds on Quest 2. This is achieved by removing shaders and post processing, and replacing it with other lightweight techniques.


* **What is the current behavior?**
The current version uses post-processing for the ocean, underwater effects, etc. This is very computationally expensive on Android devices. The scenario often runs with frame rates lower than 20 fps as a result. This is very uncomfortable and disorienting, and also has the effect of making the camera controls much slower (as they are not independent of the frame rate). In addition, the ocean surface effect is broken, resulting in strange lines moving across the surface. See issue: #222 

* **What is the new behavior?**
The frame rate now averages at 72 fps (which is the maximum allowed by default on the Quest 2) and is generally very stable. It drops when looking towards the fish cages/ocean when outside, but still mostly stays above 50 fps. This is not very noticeable, as there is little motion in front of the player when looking towards the ocean like that. The improvement in performance even allows for more fish, so there are now 400 fish per cage (1200 in total) plus some "idle fish" while still maintaining 72 fps. 
The performance improvement is achieved by removing post-processing effects. The water surface is replaced by two planes with scrolling textures. It does not look as impressive as the shader version, but that is broken (strange artifacts and worse lighting) on Android anyway. The underwater shader is replaced by Unity's built-in fog, which looks very similar while also running much faster.  


* **Does this PR introduce a breaking change?**
No, not likely. It does however include changes in both fish feeding and fish welfare scenes, especially fish feeding.

* **Other information:**
The performance was measured using OVR Metrics Tool (https://developer.oculus.com/documentation/unity/ts-ovrmetricstool/). This allowed me to monitor frame rate, CPU and GPU usage and so on to identify where the application bottlenecked on the Quest 2. Install it as any other Android app on the Quest 2 (I used adb, but SideQuest might work as well). Open it, and enable "Persistent Overlay" to see the statistics in real time as the app is running. Enable "average FPS" to verify that the frame rate is at or close to 72 fps.