@eamathie 2023-10-24
> The app builds, but the merd camera isn't working at all, it's throwing an error when I run in editor saying "UnassignedReferenceException: The variable k__BackingField of MerdCameraController has not been assigned.".
> 
> But preformance is way better in fishfeeding now

Have you pulled all the changes in the scene? The cameras had to be put into camera array in the controller again because of a change in the hierarchy